requirements.txt
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
crontab==0.23.0
idna==3.3
mysql-connector-python==8.0.30
NBT==1.5.1
nbtlib==2.0.4
numpy==1.23.1
protobuf==3.20.1
pycodestyle==2.9.0
pycron==3.0.0
python-crontab==2.6.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.11
